--- packages/ui/src/components/Form/FormContext.tsx
@@ -1,7 +1,0 @@
-import { createStyledContext } from 'tamagui';
-
-import { BackgroundType } from './formUtils';
-
-export const FormContext = createStyledContext<{
-  backgroundType: BackgroundType;
-}>({ backgroundType: 'primary' });
